(maint) Install python3 on RHEL 8
RHEL 8 does not come with a python installation by default, so unlike on
other platforms, neither `python` nor `python3` is available as an
executable. This commit updates the test that uses python to install
python3 prior to attempting to use it.

The long term fix here is to remove Python from our test suite, see
SERVER-2501.