(maint) Require environment in v4 catalog endpoint
Previously we were not requiring the caller of the v4 catalog endpoint
to provide an environment, assuming that the caller would know when it
was needed and supply the environment in those cases.

This led to the need to create defaults, either a specific environment
like "production" or Puppet's ":current_environment" within its context.
Neither of these are good defaults as default environments could be set
by users to some value besides production and
Puppet.lookup(:current_environment) has undefined, non-thread safe
semantics.

The simplest and most backwards compatible behavior is to always require
and environment from the endpoint caller.

This may be loosened in the future if we can define more concrete
workflows that may break backwards compatibility.