(maint) Puppetserver shouldn't include facter 2
Due to dependencies of the puppetserver-ca gem, the puppetserver package
includes the facter2 gem. This can lead to strange behavior in some
cases. This commit updates the mri gem installation to run with
`--no-dependencies` to not pull in facter2. This will break if that gem
ends up with additional dependencies, or if puppetserver ever removes
its dependency on the puppet-agent.