(SERVER-2532) Honor env cache refactor in etag check middleware
Previously, we updated the EnvironmentClassCache to simply be an
EnvironmentCache with tags keyed by the service they represent. However
our middleware that checked for etags was still always checking the
class info service cache.

This updates the etag check middleware to discern the required info
service from the incoming request and check the etag vs the appropriate
tag.