(SERVER-2532) Fold update-env.-info-entry into callers
Previously we used a single function for updating the environment info
cache, sometimes passing nil into that fn to invalidate its contents.

This is problematic because invalidating the cache needs to update all
info service tags, but normal updating a tag should only update the tag
for a single info service.

To resolve the mismatch between needing to update one or more tags, this
folds the tag updates into the calling methods.