(SERVER-2532) Add rsapi to modules; enable approp. tests
This removes our current dummy data returned by the
getTransportsForEnvironment interface method in favor of a basic
implementation based on a vendored version of the Resource API.

The current Resource API does not register schemas in a way that is
persistent process (reload on update) safe or cache aware. The tests
currently work around these limitations.

Our current assumption is that the Resource API will be updated to fix
these limitations and provide a better list interface which, when
taken up by Puppet Server will require removing most of the current
getTransportsForEnvironment impl and updating the integration tests.