(SERVER-2532) Enforce ordering of transports response
Our environment_transports endpoint returns a JSON object that contain
information retrieved from JRuby. It also provides etags created from
the checksummed value of the JSON string. Consequently the order of the
JSON object members and the map keypairs that they're created from
matters.

We have not see an issue in normal usage because the underlying
data structures that JRuby and Clojure use are ordered maps upto at
least eight members and all transport schema connection_infos created so
far falls under that eight member limit.

This patch forces the return map that is serialized by the transport
endpoint to be ordered by the same fn that ensures this for the classes
endpoint.