(SERVER-2532) Set :cache-generation-id per info service as :version
Previously, we had one "cache-generation-id for each environment roughly
representing the version of the environment when tags for the info
services where created. However, since saving a tag for each info
service moves the generation-id forward this caused a race condition, eg
if the classes info was being generated with a generation-id of 2 and
during that time the transports tag was saved and the generation-id
moved to 3 when we attempted to save the classes tag it would be denied
because the generation-id had moved changed underneath it.

This creates a per info service version to prevent these race conditions
between them. This "version" represents the version of the content that
the tag was generated from, but versions of content between info
services are not comparable.