(SERVER-2532) Work around non-reloadable transport schemas
In implementing this feature it was discovered that the caching
implemented in the RSAPI prevents reloading changed network transport
schemas when the environment-cache API is called. This is a change that
must occur within the RSAPI and until this is fixed our current means of
testing caching will not work.

Consequently, this commit moves all of our testing to mock out our
transport values at the JRubyPuppet protocol level. This commit should
be reverted or otherwise replaced when the bug in RSAPI is resolved
(see FM-8092).