(SERVER-2532) Reword caching symbols and docs
The way the EnvironmentClassInfoCache was implemented and documented often
confused those who first came to the project, specifically that it was
caching Etags rather than domain values and that the cache was checked
in middleware away from the explicit cache handling code in the request
handler.

During this work the purpose and interface for the EnvironmentCache (nee
EnvironmentClassInfoCache) change somewhat and a small effort was made
to update it and its consumers documentation. That effort largely
degraded coherence and understandability of the documentation and code.

This commit is an attempt to go systematically from the
jruby-puppet-core functions and documentation and update variables,
functions, and documentation in it and it's callers to better describe
what is happening, where, and why.