(SERVER-2576) Default to including all new and old ciphers
Previously, we depended on the default values of the ciphers in
tk-jetty9. With tk-jetty9 3.0 it includes a very up to date list of only
strong ciphers. This creates a default value for the cipher-suites
configuration option that includes all valid cipher suites old and new.
This gives users backwards compatibility with old ciphers (though jetty
will warn about them being weak) as well as access to the newer stronger
ciphers.