(maint) Don't override leiningen's default repositories by default
With the current repository setup in `project.clj`, `:repositories` and
`:deploy-repositories` are set to point to internal-to-puppet
infrastructure. While builds still succeed with these settings on
external networks, there is a noticible slowdown while waiting for
network timeouts.

This commit also adds an `internal` profile that can be used for builds
running on Puppet's internal network to access local repos/mirrors. To
use this new profile, run `lein with-profile ezbake,internal ezbake
build` in instances where you are currently running `lein with-profile
ezbake ezbake build`.