(maint) Treat tk-ws-jetty9 as a standard dependency
Previously we did not declare tk-jetty9 as a direct dependency. This has
existed since the very first commit to puppetserver so I'm unclear
exactly why this is but I think it is so consumers of Puppet Server
could substitute out a different webserver library that adhered to the
same tk protocol (tk-ws-jetty7?).

In practice we don't allow switching out Jetty versions in our product,
have left the only alternative to tk-jetty9 that we own unmaintained,
and, I am at least, unaware of any consumers who might have built their
own webserver libraries that they are substituting out when building
Puppet Server themselves.

Consequently, this removes the specification of tk-jetty9 from the dev
and uberjar profiles and moves it into our normal dependencies.