SERVER-2597: Set HTTP basic auth on GET in HttpClient
By moving the credential code handling from post() to
create_common_request_options() the HttpClient will handle HTTP basic
auth on get() as well.

Any code that already uses HttpClient (such as Hiera lookup functions)
and manually sets Authorization as a workaround, would start to fail. To
handle this, the header values are compared and an exception is only
raised when the values differ.