(maint) Update min-lein-version and remove unneeded nil versions
Up to lein version 2.8.0 there was a bug in which parent project managed
dependencies would not be included for dependencies specified in
profiles unless they included a nil value for their version.

This upgrades us to use the version of lein we test with in travis-ci
which includes the fix to the nil version problem. It then removes the
now unnecessary nils.