(maint) Upgrade lein to remove nrepl dependency conflict
Lein includes nrepl in its base profile. This causes conflicts with tk's
nrepl dependency and gives unclear errors when that nrepl comes in via
parent projects. Because of this, we explicitly depended on tools.nrepl
to resolve this conflict.

However, from 2.8 -> 2.9 lein upgraded from tools.nrepl 0.2.13 to
nrepl/nrepl 0.6.0. In tk 2.0.0 we also upgraded to nrepl/nrepl 0.6.0.
Consequently, when using lein 2.9+ we don't have a conflict with tk's
nrepl because they are both nrep/nrepl 0.6.0.

The next time we need to resolve a nrepl conflict with lein would be
with nrepl/nrepl rather than with tools.nrepl so we remove the
tools.nrepl dependency completely.