(SERVER-1914) Replace ssl_context w/in Puppet's context
Puppet's context contains a proc that when called will initialize an
SSLContext from Puppet's certificate infrastructure. We cannot run that
code in a JRuby FIPS environment. We also never use Puppet's SSLContext
because we manage our own.

However, Puppet's report processor will retrieve the SSLContext from the
Context to give to Puppet::Network::HttpPool.connection, which will
validate that there is some value for the SSLContext before handing off
to the HttpClient.

This places a value in the Context that the report processor may
retrieve, will pass the HttpPool's existence check, but will not trigger
Puppet's default SSLContext creation proc.