Allow request handling with one shared JRuby instance
This commit allows you to specify "jruby-puppet.multithreaded=true" to
switch into a mode where all requests requiring a jruby instance are
passed through a single JRuby, designed to be running a thread-safe
puppet implementation.

It does not address other uses of JRuby instances (like settings
loading), it does not respect `max-requests-per-instance`, and it does
not work with pool locking (the instance is never returned so a request
to lock will block forever).