(SERVER-2543) Add endpoints for plans list and plan metadata
This adds the endpoints to expose the plan information service
from Puppet as an HTTP service.

Primarily this code is copy&paste from the similar glue for tasks,
with bits removed as needed (mostly around metadata and file lists,
since currently plans have neither of those). It seemed better to
duplicate rather than try to share code paths, since task and plan
metadata formats are more likely to diverge than converge.

The test for this currently includes an assertion that plan metadata
is empty, which will be a nice reminder for us to add tests for
metadata when it becomes available :sunglasses: