(maint) Remove jira-ruby gem
The jira-ruby gem used to be used for generating tickets via a rake
task, but the only tickets we autogenerate now come from a centralized
repo of tasks, so we don't need this dep anymore. This commit removes
it, following an update to it that started limiting the Ruby versions
with which it can be installed.