(maint) Use nightlies repo to install dependencies at build time
With
https://github.com/puppetlabs/puppetserver/pull/2145#issuecomment-521705769,
puppetserver added a dependency on a not-yet-released puppet-agent. This
update uses the repos at nightlies.puppet.com instead of apt.puppet.com
for resolving the agent runtime dependency.

This change should prevent us from ending up in this situation in the
future, though also comes with some potential to destabilize things in
the containers.