(maint) Update Puppet submodule
Note this does not update the version we acceptance test with. We will
bring that in during a normal "last passing" bump. However, this
submodule update will unblock testing - the failures in question
only happen during jruby integration test and some long lived PE tests,
which are unaffected by this update anyways.