(maint) Push context changes globally during Server init
Puppet @ a1153bf3b created a method to update the default
Puppet::Context that all request threads see. We need to put in a dummy
value for the :ssl_context in the Puppet::Context during Server init to
function properly in FIPS installs.

This updates us to take advantage of that global update method, fixing a
regression in FIPS functionality.