(SERVER-2654) Remove cipher-suite config tests
These tests, while valuable when plumbing through the Java/Ruby interop
are testing basic configurability of the underlying Java library and are
better suited (and duplicated) in its tests.

The use cases are also problematic with changing supported defaults and
FIPS support. This patch removes the tests and defers the testing of
the http client libraries configuration to the library itself.