(SERVER-2654) Do not test invalid keylength on FIPS
FIPS has just as strict requirements on keylengths as non-FIPS, however
it also goes into an error state that causes all other tests to fail.
These tests consequently pollute the global state for other tests.

Ideally we would validate the keylength before passing them through to
the ssl-util library. However, that would best be done in a
`validate-settings!` and is a refacter we might not want to do at this
time.