(maint) Only build a release build if it makes sense
In our github actions configuration, we always want to try to make a release
build so that we don't have to kick that off manually. But we also only
want to build and publish a release build if it doesn't already exists.

This adds some safeguards so we don't try to build the release build yet
if it hasn't been published and updated in dujour, and also so that we
only build and publish the release build if we haven't already pushed
that tag.