(SERVER-2684) Take up new jruby-utils
This commit pulls in the new version of jruby-utils, which contains an
optional new backend that hands out references to a single JRuby
instance, instead of maintaining a pool of instances. This can be
enabled by setting `jruby-puppet.multithreaded` to `true`.

Note that the code running inside the JRuby instance must be
threadsafe to work properly. It is not recommended to enable
this mode at this time, as Puppet is not yet fully threadsafe.

The new version of jruby-utils requires that the pool-context be passed
when returning an instance to the pool, as part of the changes to
support the multithreaded Reference Pool. So this commit also updates
our use of that API.