(SERVER-2696) Set connection state to close on queue limit hits
This commit updates the 503 response returned to Puppet Agents when the
max-queued-requests limit is hit to include a "Connection: close"
header if the "Retry-After" header is used. This will cause the Puppet
agents to close their TCP connections before sleeping for the amount of
time specified by "Retry-After" which reduces load on the server from
dead connections.