(maint) Remove test cases that have bad behavior with mock JRubies
The tests for the (deprecated) whitelist access control for the admin
API were previously testing each of the various endpoints under
`puppet-admin-api`, even though the whitelist applies the same to each
of them. Moreover, two of the endpoints (to get thread dumps and flush
the JRuby pool) also do not behave correctly with mock JRubies.

Since the whitelist tests only need to show that the endpoint returns a
success code instead of 'access denied', this commit updates them to
only test one of the endpoints, the one that behaves correctly when
JRubies are mocked, as a representative case for testing the whitelist.
The other two endpoints are tested for their functionality elsewhere.