(SERVER-2680) Configure `multithreaded` in tests via env var
This commit adds a pre-populated value for the `multithreaded` setting
to the tests configs. It is false by default, and can be turned on by
setting `MULTITHREADED=true` in the environment.

When multithreaded mode is turned on, only the tests in the
`:multithreaded` test filter should be run, as the excluded tests will
fail in this mode.