(SERVER-2726) Implement thread aware Puppetserver settings
This implementation of a Puppetserver puppet settings class wraps
the original Puppet::Settings and generates new Puppet::Settings
as necessary per compilation. If, over the course of a compilation,
the settings are modified, there will be a new ThreadLocal settings
object that will be used for the duration of that compilation.
Those modified settings will not persist after the request has
been handled.