Allow executables in Docker entrypoint; avoid chmod; fix run ordering
Because the script is run with `set -e` executing `chmod` on a read-only file
will make the script exit. With these changes `chmod` is only run for files that
end with `.sh` and that are not already executable.

As `find` doesn't guarantee any particular ordering `sort` has been added to the
execute pipelines to ensure that executables are run in correct (numeric) order.