(SERVER-2758) Pull in jruby-utils override in ezbake profile
This commit adds the jruby-utils dependency to the list for the ezbake
profile, since we want to be able to override the version specified in
clj-parent.

In the future, we should look into a way of doing the ezbake dependency
overrides that doesn't involve replacing the whole dep list, since it
means that adding version overrides in the main dep list in project.clj
will be ignored when building ezbake packages unless explicitly
specified in that profile as well.