(SERVER-2718) Return authoriziation extensions with cert status response
This commit adds a new field to the GET response for the
`certificate_status` endpoint, listing the authorization extensions
present on the cert or CSR being queried. This will be used to help
inform users of potentially risky extensions, so they can make an
informed decision about whether or not to sign a certificate.