(PE-28877) Add option to compile-ast endpoint to load bolt types
Previously when the compile endpoint was asked to compile AST with references to Bolt objects the compile would fail. This commit updates the compiler to accept serialized pcore and to deserialize it for compilation. This adds the requirement of some Bolt library code (and some supporting gems [logging, addressable, public_suffix]) as well as the boltlib module. Part of making the serialized types perform as expected the compiler now creates specialized bolt inventory to back operations on bolt Target options. This work depends on two new additions to the `jruby-puppet` configuration section for pe-puppetserver. The first is appending the path to the bolt library code in the `ruby-load-path` array. The second is a new setting called `boltlib-path` which is an array pointing to the path of any bolt modules to prepend to the modulepath for compilation. The compile request now accepts an option to load bolt types. The default behavior is to *not* attempt to load bolt types.