(maint) Update release notes formatting through 2018
Through 2018 and most of 2019 the format of our release notes was fairly
consistent wrt bulleted lists for note content. The last few releases
don't strictly follow this and the pre-2018 docs seem to follow a fairly
different style.

This updates those recent release notes not following the 2018 and later
style to uniformily use bulleted lists for individual note items.