(maint) Remove max-queued-requests warning note
This commit removes a warning against using the `max-queued-requests`
setting with a node population that includes significant numbers of
agents running a version older than 5.3.1. The fixes for SERVER-2405,
and SERVER-2696 shipped in Puppet Server 5.3.11 and 6.9.0 make this
warning obsolete as Puppet Server now checks the agent version headers
and skips enforcement of the limit if the agent is too old or the
header is missing.