(SERVER-2859) Add an API endpoint to clean certs
This commit adds a new endpoint, `puppet-ca/v1/clean`, that can accept
one or more certnames for cleaning, that is, revoking and removing cert
files. If any cert is missing from the CA, the request will proceed and
return a body missing the certs that could not be cleaned.

It also accepts an `async` key that will later be used to add
command-style processing to the endpoint, that does not block the
client, since processing could potentially take a very long time if many
certs are requested to be cleaned.