(SERVER-2677) Safely remove ssl dir
Prior to this change, the pre-suite would nuke the ssl dir while
puppetserver was running; this would crash the puppetserver,
and systemd would then restart the service. This results in a race
condition with the beaker method #with_puppet_running_on. If the
restart from the crash completes after #with_puppet_running_on ensures
that the service is stopped, then puppetserver never picks up the new
puppet.conf file. I think...

If this ends up solving the transients, the other SERVER-2677 commits
should be reverted.