(SERVER-2780) Update HttpClient to new API
This commit updates our Ruby wrapper for the apache HTTP client to match
the new API of Puppet's HTTP client, so our Java implementation can be
swapped in when running in Puppet Server without requiring Puppet to use
an adapter class.

The most notable difference between this implementation and the old one
is that instead of providing the hostname and port when the client is
created, the whole URL must be passed into each request, along with its
protocol (http vs https). Nearly all of the spec updates in this commit
are changes to match that API change.