(SERVER-2809) Add rake task for updating facter submodule
This commit adds a new rake task that can be used to update the Facter
submodule commit based on the last passing version published by the
puppet-agent Jenkins pipeline.

It introduces a new variable to track which branch to pull the Facter
SHA from, because while the 6.x agent is still using Facter 3, in our
submodule we use Facter 4, because we need it to be a Ruby lib rather
than C++.

This also renames some tasks to be more descriptive of the new state, so
any automation around these tasks will need to be updated accordingly.