(SERVER-2780) Ensure response object contains URI
This commit updates the response object's URL field to contain a URI
object instead of just a URL string. This object is retrieved from the
Java RequestOptions, so that it is representative of the URI object that
was used while making the request, after any processing done either in
Ruby or in Java.