(SERVER-2780) Add note about ignoring some passed options
Puppet's HttpClient, on which ours is based, accepts some options to its
constructor, including `redirect_limit` and `retry_limit`. Our client
will ignore both of these, since the former isn't easily configurable
with the API we are currently using, and the latter isn't supported at
all by the Apache HttpAsyncClient in the version (4.x) we are using.