(SERVER-2780) Add user-agent header to match Puppet's default
This commit adds a default "User-Agent" header, based on Puppet's
`http_user_agent` setting. It also adds a note about not supplying an
"X-Puppet-Version" header. Eventually we should probably provide the
puppetserver version as a header, but the work to plumb that value
through to the Ruby code is outside the scope of this PR.