(maint) Stop registering server HttpClient class
We want to defer to Puppet's shim implementation for the HTTP client
class registered with the HTTP pool. This shim will adapt calls made to
the old-style HTTP client API to the new form, which will then get
passed to our client impl, which is registered as
`Puppet.runtime[:http]`.