(SERVER-2760) Symlink new to old cadir on bootstrap
With Puppet 7 we want to move the cadir outside of the ssldir however we
want to symlink back to the old cadir to ease the transition. This
provide existing documentation and code to largely just work for new
users, preventing them from being able to accidentally delete their
cadir.

This commit looks at the last portion of the cadir to determine if it is
the the new default or something else while also allowing us to
substitute out the confdir/root for easy testing and switching between
running puppetserver as a regular user or as root. If the last portion
of the cadir matches the new default we will symlink back to the old
cadir location. Otherwise nothing happens.