(SERVER-2870) Prefer `server` to `master` in Puppet APIs
Puppet has updated its APIs and config sections to deprecate the term
`master` and prefer `server` instead. This updates us to use the term
`server` where it would trigger a deprecation warning for users during
normal operation. This is a partial implementation of SERVER-2870 as it
only targets those APIs that cause deprecation warnings.