(SERVER-2913) Use Jetty default ciphers
In Server 6.5, Jetty began warning when using what it considered weak
cipher suites. Since then, we've also audited the cipher suites for
FIPS compliance to only support modern cipher suites.

In FOSS however, we've shipped with the older cipher suites that were
enabled at the start of the 6.x series.

With Puppet 7, remove the older ciphers and bring FOSS up to PE
standards.