(maint) Fix decode failure test case
One of the updates to the ring-codec library made its `percent-decode`
function smarter, so it only detects `%` with two alphanumeric
characters after it as encoded bytes. Our test case used a string with
`%` followed by only one other character, so it didn't trigger the
"decode". This commit updates the test to use a string that looks like a
potentially valid encoding.