(maint) Remove unused metrics helpers
These helpers are leftover from before all the metrics testing moved to
the PE suite. They are duplicated in the helpers there. However, due to
Ruby loadpath shenanigans, they were still getting used sometimes
instead of the versions in PE. This is misleading and unnecessary, so
this commit removes the unused FOSS versions.