(SERVER-2948) Treat ASCII-8BIT body as bytes
Puppet returns binary data for some endpoints as a string encoded with
Ruby's ASCII-8BIT encoding. This encoding is unique to ruby and is
essentially meant to be treated as raw bytes. As of version 1.6.0,
ring-servlet attempts to encode the response body when it's a string
using the charset in the header, falling back to UTF-8. When we
encounter this encoding now we will convert it to a java byte array and
ring won't attempt to encode it.