(PE-30976) Pin protobuf test dep to 3.4.0
One of the libraries we use in testing, cemerick/url, is unmaintained
and archived. It has not been updated to pull in a non-vulnerable
version of protobuf, so this commit overrides that dependency to pull in
the fixed version.

Ultimately we should probably try to stop using cemerick/url. When that
happens, we can remove this pin as well.