(SERVER-2338) Always sign agent certs with CN as SAN
Before this change, csr's without specified alternative names were
not given one at signing; this behavior was deprecated in RFC
2818, https://tools.ietf.org/html/rfc2818#section-3.1. Many TLS
clients, now including Chrome, have specifically removed this
support.

This change ensures that when signing an agent cert, the CN in the
CSR is always added as a SAN to the resulting certificate.