(SERVER-2558) Use puppet's copy of CRL as jetty CRL
This commit changes the default CRL path supplied to Jetty from the CA's
to the local server's copy. This latter path, in the ssldir rather than
the CA, is present even if the CA service is disabled, because the
puppet agent will write the CRL it fetches to that location, while the
CA's copy is not always available.

This mirrors what we configure in PE, and makes setting up compilers in
FOSS easier.