(maint) Add single catalog compile script
This commit puts the original thread test script into a `parallel_test`
script, and creates a new test script that just makes a single catalog
request. The former is good for testing races, while the latter is
better for inspecting an individual compile.

It also updates the fixtures to no longer use the `future_features`
setting, which has been removed.