(SERVER-2867) Make server-* configs preferred to master-* variants
This creates new server-* variants of the existing master-* config
names (master-conf-dir, master-code-dir, etc). When resolving values
we will prefer the new setting, the old setting, then the default.

This updates the default config files shipped with Puppet Server to use
the new setting names. It also updates the variable names containing the
default values in code. This updates the tests to use the new names,
changes the name of the variables that hold their test values, and in
select cases changes the names of the test directories from `master` to
`server` (only for the autogenerated test directiories, not the
fixtures).