(maint) Make server-* configs available at master-*
When we moved to prefering server-* config variants we did not set
master-* variants if the server-* variants were set first. This will
cause a regression in those who programmatically use our config.

This should only cause issues for those who load puppetserver as a
library, which may only be pe-puppetserver-extensions.

This patch resolves server-* variants as usual and then sets the
master-* variants to the equivalent server-* values.