(SERVER-3014) Stop putting double slashes in test urls
As of version 4.5.8 of the apache httpclient, URLs with a double slash
in them are processed differently. Our tests were accidentally adding an
extra slash between the server and the URL path, which ran afoul of
this. This commit updates some callers of the test helper to stop adding
an additional slash.