(SERVER-3015) Cache etag if cache was previously evicted
Prior to this patch, if a request to the environment_classes endpoint contained
an If-None-Match header with an etag in it, but the cache had been evicted we
would compute the class info. If that class info's checksum matched the given etag
we would return a 304 but not cache the checksum. Because the cache was not
updated, subsequent requests that included an etag would still trigger class info
computation.

With this patch we will now cache the computed etag, allowing subsequent
requests that include the correct If-None-Match header to skip class info
computation.