(SERVER-2550) Don't warn about not replacing our CA CRL
This commit moves the logic to avoid allowing our CA CRL to be replaced
out of the replacement loop, to avoid visiting it every time this
endpoint is hit, regardless of what CRLs are sent. It also removes the
logging when we skip it. That detail would be better called out in docs.