(PE-32070) Update Java security files for new BCFIPS
There is an issue with bc-fips 1.0.2.1 involving circular provider
loading, that can be fixed by using the full class name for the
BCJSSE FIPS provider, rather than relying on an alias. This commit
updates our java-security files to use the full name. Note this only
caused issues on Java 11, but it is valid for both versions, so we are
updating it in both files for consistency.

It also updates the key manager factory algorithm following a
recommendation from the BC maintainers.