(SERVER-3032) Always allow SAN that matches CSR subject
Previously, when `allow-subject-alt-names` was set to false, we rejected
any and all SANs on CSRs. However, now that we are adding the CSR's
subject as an alt name when generating CSRs, per RFC 2818, we need to
make an exception, so as to not block all internally generated CSRs.
This affects, for example, the `puppetserver ca generate` command. This
commit updates the validation function to allow a single alt name
matching the CSR's subject, even when `allow-subject-alt-names` is
false.