(SERVER-3058) Add "scripts" mount as a default mount
This adds a new default mount "scripts" which loads files from the
scripts directory of the specified module. The mount can be used by the
file_content and file_metadata APIs (or any other API that specifies a
mount) to load file metadata and content for scripts from modules.

Requires puppetlabs/puppet#8711