(SERVER-2509) Prevent revoked cert to be add to CRL
This commit implement a filter that apply to the list of serials
number to be added to the revoked list of Puppet's CRL.  The purpose
of the filter is to remove any serial number of the certificates that
are already in the CRL's revoked list. This filter is to be used
with the 'revoke-existing-certs!' function so that it prevents the case
of having an extremely large CRL file.