(SERVER-3020) Improve certificate_status endpoint speed for querying
This commit implements a new function 'get-csr-or-cert-statuses'
that get the statuses for the queried state by the user. Before
this commit, the behavior of querying for CSRs statuses is to
get the statuses for all signed certificates as well as CSRs and
then apply a filter to get the desired state. This commits made
it so that when the user wants to query CSRs, only the directory
of CSRs will be walked through. By doing this, the speed for
querying CSRs will improve.