(maint) Provide a reasonable default for bolt-projects dir
For some testing it is desireable for puppetserver to serve some project content from some pre-defined project content. By setting a bolt-projects dir we can enable puppetserver to serve project content assuming it is in a proper structure even if it is not actually managed with code manager/file sync.